// src/jwt/types.ts
export type Key = jose.KeyLike | Uint8Array;
// ...
export const EC_ALGORITHMS = ["ES256", "ES384", "ES512"] as const;
export const RSA_ALGORITHMS = ["RS256", "RS384", "RS512"] as const;
export const HMAC_ALGORITHMS = ["HS256", "HS384", "HS512"] as const;

export const ALGORITHMS = [...EC_ALGORITHMS, ...RSA_ALGORITHMS, ...HMAC_ALGORITHMS] as const;

export type ECAlgorithm = typeof EC_ALGORITHMS[number];
export type RSAAlgorithm = typeof RSA_ALGORITHMS[number];
export type HMACAlgorithm = typeof HMAC_ALGORITHMS[number];

export type Algorithm = typeof ALGORITHMS[number];
// ...
export interface JWTPayloadOptions {
    readonly iss: string | undefined;
    readonly sub: string | undefined;
    readonly aud: string;
    readonly iat: Date | undefined;
    readonly nbf: Date | undefined;
    readonly jti: string | undefined;
}
export interface JWTPayloadOptionsDefault extends JWTPayloadOptions {}
// ...
export interface IssuerOptions<O extends JWTPayloadOptions, P extends {} | undefined, H extends {} | undefined> {
    readonly additionalHeaders: H;
    readonly algorithm: Algorithm;
    readonly validTimeMs: number;
    readonly validityDelayMs: number;
    readonly nameOrUrl: O["iss"];
    readonly generateJWTID: () => O["jti"];
    readonly defaultPayload: P;
}

export interface IssueTokenOptions<O extends JWTPayloadOptions, P extends {} | undefined, H extends {} | undefined> {
    readonly audience: O["aud"][] | O["aud"] | undefined;
    readonly validTimeMsOverride?: number;
    readonly validityDelayMsOverride?: number;
    readonly subject: O["sub"];
    readonly additionalPayload: Partial<P>;
}

// ! Token manipulation interfaces
/**
 * A class that implements the Parser interface is capable of parsing tokens from their string from into their object form.
 */
export interface Parser<T> {
    /**
     * Parses a token from its string form into an object form
     * @param token The string representation of the JWT token
     */
    parseToken(token: string): Promise<T>;
}
/**
 * A class that implements the Verifier interface can verify a tokens signature.
 * A class that implements the Verifier interface also needs to implement the Parser interface.
 */
export interface Verifier<T> extends Parser<T> {
    /**
     * Verifies the token in its object form
     * @param token A token in an object form
     */
    verifyToken(token: T): Promise<boolean>;
    /**
     * Parses a token and then verifies its signature
     * @param token A token in a string form
     */
    parseThenVerifyToken(token: string): Promise<T | false>;
}

/**
 * A class that implements the Issuer interface is capable of issuing tokens.
 */
export interface Issuer<T, O> extends Verifier<T> {
    /**
     * Issues a token based on provided options
     * @param options Options for the token
     */
    issueToken(options: O): Promise<T>;
}